UI changes to add new "Edit plugin settings" section for extensions
