+++ client/my-sites/plugins/plugin-meta/style.scss
@@ -152,23 +152,8 @@
 	color: $alert-green;
 }
 
-.plugin-meta__settings-link {
-	clear: both;
-	padding-top: 24px;
-	text-transform: uppercase;
-}
-
-.plugin-meta__action-links {
-	display: flex;
-	flex-wrap: wrap;
-	clear: both;
-	padding-top: 12px;
-
-	.button {
-		margin-top: 12px;
-		flex-shrink: 0;
-		margin-right: 10px;
-	}
+.card.plugin-meta__plugin-information-wrapper {
+	margin-top: 16px;
 }
 
 a.plugin-meta__settings-link {
